quick commit, see next commit for details
