+++ maserol/core.py
@@ -588,27 +588,3 @@ def assemble_bounds(data, rcps, logistic_ligands):
     )
     assert bounds[0].size == bounds[1].size
     return bounds
-
-
-def Rtot_to_xarray(Rtot: np.ndarray, data: xr.DataArray, rcps: List):
-    return xr.DataArray(
-        Rtot,
-        coords={
-            "Complex": data.Complex,
-            "Receptor": rcps,
-        },
-        dims=["Complex", "Receptor"],
-    )
-
-
-def Rtot_to_df(
-    Rtot: Union[xr.DataArray, np.ndarray], data: xr.DataArray = None, rcps: List = None
-):
-    if isinstance(Rtot, np.ndarray):
-        assert data is not None, "data required if Rtot is np array"
-        assert rcps is not None, "rcps required if Rtot is np array"
-        Rtot = Rtot_to_xarray(Rtot, data, rcps)
-    Rtot_df = Rtot.to_dataframe(name="Abundance").drop(columns=["Antigen", "Sample"])
-    Rtot_df = Rtot_df.reset_index(level="Receptor").pivot(columns="Receptor")
-    Rtot_df.columns = [col[1] for col in Rtot_df.columns]
-    return Rtot_df
