+++ maserol/figures/figure_3d.py
@@ -1,52 +1,52 @@
 import functools
 import numpy as np
 import pandas as pd
-import seaborn as sns
 from tensordata.zohar import data as zohar
 
 from maserol.impute import (
+    impute_missing_cp,
     impute_missing_ms,
+    impute_missing_pca,
     run_repeated_imputation,
 )
-from maserol.preprocess import assemble_options, get_kaplonek_mgh_data
+from maserol.preprocess import prepare_data
 from maserol.figures.common import getSetup
 
 
-LIGS = ["FcR2A", "FcR2B", "FcR3A", "FcR3B"]
-RUNS = 3
-MISSINGNESS = 0.2
+# MS 100% missing ligand vs CP and PCA 95% missing
+LIGS = ["FcR2A"]
 
 
 def makeFigure():
-    tensor = get_kaplonek_mgh_data()
-    tensor = tensor[np.random.choice(tensor.shape[0], 300)]
-
-    fs = [1, 4]
-    optss = [
-        assemble_options(tensor, ["IgG1", "IgG1f", "IgG3", "IgG3f"], FcR_f=f)
-        for f in fs
+    tensor_3d_zohar = (
+        zohar()
+        .rename({"Receptor": "Ligand"})
+        .sel(Ligand=["IgG1", "IgG3", "FcR2A", "FcR2B", "FcR3A", "FcR3B"])
+        .transpose("Sample", "Ligand", "Antigen")
+    )
+    tensor_3d_zohar_sub = tensor_3d_zohar[
+        np.random.choice(tensor_3d_zohar.shape[0], 200)
     ]
-    for opts in optss:
-        opts["tol"] = 1e-6
-    imputers = [functools.partial(impute_missing_ms, opts=opts) for opts in optss]
+    tensor_2d_zohar_sub = prepare_data(
+        tensor_3d_zohar_sub.rename({"Ligand": "Receptor"})
+    )
+    imputers = [
+        impute_missing_ms,
+        functools.partial(impute_missing_cp, ncomp=5),
+        functools.partial(impute_missing_pca, ncomp=5),
+    ]
+    tensors = [tensor_2d_zohar_sub, tensor_3d_zohar_sub, tensor_2d_zohar_sub]
+    missingnesss = [1, 0.95, 0.95]
+    runss = [1, 3, 3]
     df = pd.concat(
         run_repeated_imputation(
-            tensor,
-            imputer,
-            ligs=LIGS,
-            runs=RUNS,
-            missingness=MISSINGNESS,
-            imputer_name=f,
+            tensor, imputer, ligs=LIGS, runs=runs, missingness=missingness
         )
-        for imputer, f in zip(imputers, fs)
+        for tensor, imputer, missingness, runs in zip(
+            tensors, imputers, missingnesss, runss
+        )
     )
-    axes, fig = getSetup((14, 6), (1, 2))
-    sns.boxplot(data=df, x="Ligand", y="r2", hue="Method", ax=axes[0])
+    axes, fig = getSetup((7, 6), (1, 1))
     axes[0].set_xlabel("Ligand")
-    axes[0].set_ylabel("Coefficient of Determination")
-    axes[0].legend(title="Valency")
-    sns.boxplot(data=df, x="Ligand", y="r", hue="Method", ax=axes[1])
-    axes[1].set_xlabel("Ligand")
-    axes[1].set_ylabel("Pearson Correlation")
-    axes[1].legend(title="Valency")
+    axes[0].set_ylabel("Pearson Correlation")
     return fig
