+++ maserol/test/test_preprocess.py
@@ -1,7 +1,6 @@
 import pytest
 from tensordata.atyeo import data as atyeo
 from tensordata.zohar import data as zohar
-from tensordata.kaplonek import MGH4D, SpaceX4D
 from ..preprocess import prepare_data
