+++ pyproject.toml
@@ -17,12 +17,14 @@ scikit-learn = "^1.1.2"
 PyYAML = ">=6.0"
 statsmodels = "^0.14.0"
 tensorly = "^0.8.1"
+statannotations = "^0.6.0"
 
 [tool.poetry.group.dev.dependencies]
 pytest = "^6.2"
 pytest-cov = "^3.0"
 mypy = "^0.991"
 valentbind = {git = "https://github.com/meyer-lab/valentBind.git", branch = "main"}
+ipykernel = "^6.26.0"
 
 [build-system]
 requires = ["poetry-core>=1.0.0"]
